Btrfs: fix the boundary problem of the generation of the free space inode
There are two boundary problems of the generation of the free space inode:
- the address of the generation may be in the end of the first page, in this
  case, we must map the next page after we get/set the generation. Or we would
  access a invalid address when we read/write the free space entry.
- it is possible that the generation is fitted at the beginning of the second
  page, so we need skill the process of the first page.

Limiting the size of the free space inode can fix the problems when we write
out the cache, but if we load a old free space file, the above problems still
exist. And besides that, considering the flexibility of the free space cache,
it is also better that we fix the problems in the module of the free space
cache.

So we modify io_ctl_{set, check}_generation() to fix the above problems:
- Check the first page is full of CRCs or not, if yes, we map the second
  page and then get/set the generation. If it is not full of CRCs, that
  is to say the generation is fitted in it, we get/set the generation
  directly.
- After we get/set the generation, we check the pointer is at the end of
  the current page or not. If yes, we map the next page.

Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>