Btrfs: cleanup the code of the inode rb-tree operation
Before applying this patch, we erase the inode from the inode rb-tree of the
fs/file tree after we remove the inode from the hlist of the filesystem. It
caused a problem that we may touch a inode whose inode-id was the same with
the one we inserted but is under freeing. It is because when we can not find
the expected inode in the hlist of the filesystem, we will allocate a new one
and insert it into the hlist and the inode rb-tree of the fs/file tree, but
at this time, the old inode is still in the rb-tree. In order to avoid this
problem, we erase the old one and then try to insert again. It was a bit
complex.

This patch move the inode_tree_del() from btrfs_destroy_inode() to
btrfs_evict_inode(), By this way, we remove the inode from the rb-tree before
we remove it from the hlist of the filesystem, so we will not find a inode
with the same id in the rb-tree when we insert a new inode, and the code of
inode_tree_add() can become simpler and clearer.

Note: with this patch, we can not access the fs/file tree in
btrfs_destroy_inode() because the tree may be freed.

Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>