Btrfs: scan all the devices and build the fs device list by btrfs's self
The original code need scan the devices and build the fs device list by the user
tool by udev or users' selves. It is flexible. But if someone re-install the filesystem
module, and forget to scan the devices by himself, the filesystem would report that
"can not open some device" when mounting the filesystem, it was uncomfortable, this
patch fixes this problem by scanning the devices when we install the module.

Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>