Btrfs: fix the problem that device clone didn't consider the missing device
The device clone function didn't consider the missing device, and didn't
set the missing bool variant and didn't increase the missing device counter,
it would introduce some problem if someone uses the missing device counter
or missing mark of the device.

Fix it by checking the missing bool variant when cloning the device, and
because only missing device doesn't have its own name, so add a assert
for the device which is not missing device. Since we set the missing mark
for the missing device, we needn't check the name pointer after the missing
mark check when showing the device name.

Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>