Btrfs: Separate finding-right-mirror and writting-to-target's process in scrub_handle_errored_block()
In corrent code, code of finding-right-mirror and writting-to-target
are mixed in logic, if we find a right mirror but failed in writting
to target, it will treat as "hadn't found right block", and fill the
target with sblock_bad.

Actually, "failed in writting to target" does not mean "source
block is wrong", this patch separate above two condition in logic,
and so some cleanup to make code clean.

Signed-off-by: Zhao Lei <zhaolei@cn.fujitsu.com>
Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>